Improve handling of UncaughtExceptionHandler
