+++ statsnode/src/main/java/bisq/statistics/StatisticsMain.java
@@ -22,7 +22,6 @@ import bisq.core.app.misc.ModuleForAppWithP2p;
 
 import bisq.common.UserThread;
 import bisq.common.app.AppModule;
-import bisq.common.setup.CommonSetup;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -45,7 +44,6 @@ public class StatisticsMain extends ExecutableForAppWithP2p {
         super.doExecute();
 
         checkMemory(config, this);
-        CommonSetup.setupUncaughtExceptionHandler(this);
 
         keepRunning();
     }
